telegram-bot/requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
